Move table initialisations from main.py to /models
