--- backend/models/itemModel.py
@@ -1,6 +1,0 @@
-class Item:
-    def __init__(self, name, price, description):
-        self.name = name
-        self.price = price
-        self.description = description
-
